requirements.txt
Pillow>=10.0.0
PyQt6>=6.6.0
opencv-python>=4.8.0
numpy>=1.24.0
